Remove . after update npm command to avoid confusion
Being a lazy sort of person, I copied and pasted the update npm command from this page, and the trailing full-stop/period got copied too. This made npm ignore the global flag and it installed it's update in the working directory that I was currently in.

I'm suggesting removing that trailing full-stop/period to prevent this happening.